% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-mortfit.R
\name{AC_from}
\alias{AC_from}
\title{Resolve an AC family specification to a PAC object}
\usage{
AC_from(AC = "Beard")
}
\arguments{
\item{AC}{a \code{PAC} object, a family name, or a family constructor}
}
\description{
Resolve an AC family specification to a PAC object
}
\keyword{internal}
