% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-fns.R
\name{Beard}
\alias{Beard}
\alias{dbeard}
\alias{pbeard}
\alias{qbeard}
\alias{rbeard}
\title{The Beard distribution
Density, distribution function, quantile function and random generation for the Beard distribution.}
\usage{
dbeard(x, par)

pbeard(x, par)

qbeard(p, par, lower.tail = TRUE)

rbeard(n, par)
}
\arguments{
\item{x, q}{numeric vector of quantiles.}

\item{par}{numeric length-3 vector of parameters \code{c(a, b, gam)}.}

\item{p}{numeric vector of probabilities.}

\item{lower.tail}{logical; if \code{FALSE}, probabilities are \eqn{P(X > x)}.}

\item{n}{number of observations.}
}
\value{
\itemize{
\item \code{dbeard()}: density.
\item \code{pbeard()}: distribution function.
\item \code{qbeard()}: quantiles.
\item \code{rbeard()}: random values.
}
}
\description{
The Beard distribution
Density, distribution function, quantile function and random generation for the Beard distribution.
}
\details{
The Beard distribution is defined by the hazard function \eqn{\mu(x) = \frac{\exp(a + b x)}{1 + \exp(a + \gamma + b x)}}.
The survival function is \eqn{S(x) = ((1 + \exp(a + \gamma)) / (1 + \exp(a + \gamma + b x)))^{\exp(-\gamma) / b}}.
}
\examples{
dbeard(80:100, c(-10, 0.09, -3))
pbeard(80:100, c(-10, 0.09, -3))
qbeard(seq(0, 1, by = 0.1), c(-10, 0.09, -3))
rbeard(10, c(-10, 0.09, -3))
}
