% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utils-ibs.R
\name{BS_raw}
\alias{BS_raw}
\title{Raw Brier Score contribution (single i, single age)}
\usage{
BS_raw(
  x,
  birthage,
  censoredage,
  DthIndicator,
  Sfunc,
  par,
  Surv_info,
  x_min = 80
)
}
\description{
Raw Brier Score contribution (single i, single age)
}
\keyword{internal}
