% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utils-ibs.R
\name{.empirical_from_cache}
\alias{.empirical_from_cache}
\title{Empirical survival cache retriever}
\usage{
.empirical_from_cache(key)
}
\description{
Empirical survival cache retriever
}
\keyword{internal}
