% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-utils-ibs.R
\name{empirical_survival}
\alias{empirical_survival}
\title{Compute empirical survival info for IBS weighting}
\usage{
empirical_survival(data, key, x_min = 80, x_max = 115)
}
\arguments{
\item{data}{data.frame}

\item{key}{character cache key}

\item{x_min, x_max}{integer age range (inclusive)}
}
\value{
list with survfit objects and step values
}
\description{
Compute empirical survival info for IBS weighting
}
