% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-fit.R
\name{fit,PACFit-method}
\alias{fit,PACFit-method}
\title{Fit a PAC model using GenSA (if available)}
\usage{
\S4method{fit}{PACFit}(model_fit, data_train, patience = 30, tol = 1e-06, verbose = FALSE)
}
\arguments{
\item{data_train}{data.frame with birthage,censoredage,DthIndicator}

\item{patience}{integer early-stop counter}

\item{tol}{numeric tolerance}

\item{verbose}{logical}
}
\description{
Fit a PAC model using GenSA (if available)
}
