% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-methods-loglik.R
\name{get_logllh,PAC-method}
\alias{get_logllh,PAC-method}
\title{Dataset log-likelihood extractor for PAC}
\usage{
\S4method{get_logllh}{PAC}(model, data_train, par = NULL)
}
\description{
Dataset log-likelihood extractor for PAC
}
