% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-fns.R
\name{Gompertz}
\alias{Gompertz}
\alias{dgompertz}
\alias{pgompertz}
\alias{qgompertz}
\alias{rgompertz}
\title{The Gompertz distribution}
\usage{
dgompertz(x, par)

pgompertz(x, par)

qgompertz(p, par, lower.tail = TRUE)

rgompertz(n, par)
}
\arguments{
\item{x, q}{numeric vector of quantiles.}

\item{par}{numeric length-2 vector of parameters \code{c(a, b)}.}

\item{p}{numeric vector of probabilities.}

\item{lower.tail}{logical; if \code{FALSE}, probabilities are \eqn{P(X > x)}.}

\item{n}{number of observations.}
}
\value{
\itemize{
\item \code{dgompertz()}: density.
\item \code{pgompertz()}: distribution function.
\item \code{qgompertz()}: quantiles.
\item \code{rgompertz()}: random values.
}
}
\description{
Density, distribution function, quantile function and random generation for the Gompertz distribution.
}
\details{
The Gompertz distribution is defined by the hazard function \eqn{\mu(x) = \exp(a + b x)}.
The survival function is \eqn{S(x) = \exp\{-\exp(a) (e^{b x} - 1)/b\}}.
}
\examples{
dgompertz(80:100, c(-10, 0.09))
pgompertz(80:100, c(-10, 0.09))
qgompertz(seq(0, 1, by = 0.1), c(-10, 0.09))
rgompertz(10, c(-10, 0.09))

}
\seealso{
\code{\link[=survival_gompertz]{survival_gompertz()}} given the same parameterization.

\code{\link[flexsurv:Gompertz]{flexsurv::rgompertz()}} for an implementation in \pkg{flexsurv}.
}
