% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-mortfit.R
\name{gph_fit}
\alias{gph_fit}
\title{Fit a Gompertz proportional hazards model via \pkg{flexsurv}}
\usage{
gph_fit(formula, data = NULL, data_train = NULL, data_test = NULL)
}
\arguments{
\item{formula}{A standard survival formula with covariates.}

\item{data, data_train, data_test}{See \code{baseline_fit()}.}
}
\value{
A \code{GPHModel} object.
}
\description{
Fit a Gompertz proportional hazards model via \pkg{flexsurv}
}
