% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-fns.R
\name{Kannisto}
\alias{Kannisto}
\alias{dkannisto}
\alias{pkannisto}
\alias{qkannisto}
\alias{rkannisto}
\title{The Kannisto distribution}
\usage{
dkannisto(x, par)

pkannisto(x, par)

qkannisto(p, par, lower.tail = TRUE)

rkannisto(n, par)
}
\arguments{
\item{x, q}{numeric vector of quantiles.}

\item{par}{numeric length-2 vector of parameters \code{c(a, b)}.}

\item{p}{numeric vector of probabilities.}

\item{lower.tail}{logical; if \code{FALSE}, probabilities are \eqn{P(X > x)}.}

\item{n}{number of observations.}
}
\value{
\itemize{
\item \code{dkannisto()}: density.
\item \code{pkannisto()}: distribution function.
\item \code{qkannisto()}: quantiles.
\item \code{rkannisto()}: random values.
}
}
\description{
Density, distribution function, quantile function and random generation for the Kannisto distribution.
}
\details{
The Kannisto distribution is defined by the hazard function \eqn{\mu(x) = \frac{\exp(a + b x)}{1 + \exp(a + b x)}}.
The survival function is \eqn{S(x) = ((1 + \exp(a)) / (1 + \exp(a + b x)))^{1 / b}}.
}
\examples{
dkannisto(80:100, c(-10, 0.09))
pkannisto(80:100, c(-10, 0.09))
qkannisto(seq(0, 1, by = 0.1), c(-10, 0.09))
rkannisto(10, c(-10, 0.09))
}
