% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-methods-loglik.R
\name{logllh,PAC-method}
\alias{logllh,PAC-method}
\title{Log-likelihood for PAC (vectorized over observations)}
\usage{
\S4method{logllh}{PAC}(model, par, x, r, i)
}
\arguments{
\item{model}{PAC}

\item{par}{numeric parameter vector (defaults to model@params if omitted)}

\item{x}{birth age}

\item{r}{censored/event age}

\item{i}{event indicator (1=death, 0=censored)}
}
\description{
Log-likelihood for PAC (vectorized over observations)
}
