% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-fns.R
\name{Makeham}
\alias{Makeham}
\alias{dmakeham}
\alias{pmakeham}
\alias{qmakeham}
\alias{rmakeham}
\title{The Makeham distribution}
\usage{
dmakeham(x, par)

pmakeham(x, par)

qmakeham(p, par, lower.tail = TRUE)

rmakeham(n, par)
}
\arguments{
\item{x, q}{numeric vector of quantiles.}

\item{par}{numeric length-3 vector of parameters \code{c(a, b, lambda)}.}

\item{p}{numeric vector of probabilities.}

\item{lower.tail}{logical; if \code{FALSE}, probabilities are \eqn{P(X > x)}.}

\item{n}{number of observations.}
}
\value{
\itemize{
\item \code{dmakeham()}: density.
\item \code{pmakeham()}: distribution function.
\item \code{qmakeham()}: quantiles.
\item \code{rmakeham()}: random values.
}
}
\description{
Density, distribution function, quantile function and random generation for the Makeham distribution.
}
\details{
The Makeham distribution is defined by the hazard function \eqn{\mu(x) = \exp(a + b x) + \lambda}.
The survival function is \eqn{S(x) = \exp\{-\exp(a) (e^{b x} - 1)/b - x \lambda\}}.
}
\examples{
dmakeham(80:100, c(-10, 0.09, 0.0005))
pmakeham(80:100, c(-10, 0.09, 0.0005))
qmakeham(seq(0, 1, by = 0.1), c(-10, 0.09, 0.0005))
rmakeham(10, c(-10, 0.09, 0.0005))
}
