% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-classes.R
\docType{class}
\name{PAC-class}
\alias{PAC-class}
\title{PAC and PACFit S4 Classes}
\description{
S4 classes for parametric age-cohort (PAC) survival models and fits.
}
\section{Slots}{

\describe{
\item{\code{name}}{model name}

\item{\code{params}}{numeric vector of parameters}

\item{\code{survival_}}{function(par, x) -> S(x)}

\item{\code{mu_}}{function(par, x) -> hazard at x}

\item{\code{cumhaz_}}{function(par, x, t) -> cumulative hazard from x to x+t}

\item{\code{bounds}}{list(lower=, upper=)}
}}

