% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-fns.R
\name{Perks}
\alias{Perks}
\alias{dperks}
\alias{pperks}
\alias{qperks}
\alias{rperks}
\title{The Perks distribution}
\usage{
dperks(x, par)

pperks(x, par)

qperks(p, par, lower.tail = TRUE)

rperks(n, par)
}
\arguments{
\item{x, q}{numeric vector of quantiles.}

\item{par}{numeric length-4 vector of parameters \code{c(a, b, gam, lambda)}.}

\item{p}{numeric vector of probabilities.}

\item{lower.tail}{logical; if \code{FALSE}, probabilities are \eqn{P(X > x)}.}

\item{n}{number of observations.}
}
\value{
\itemize{
\item \code{dperks()}: density.
\item \code{pperks()}: distribution function.
\item \code{qperks()}: quantiles.
\item \code{rperks()}: random values.
}
}
\description{
Density, distribution function, quantile function and random generation for the Perks distribution.
}
\details{
The Perks distribution is defined by the hazard function \eqn{\mu(x) = \frac{\exp(a + b x)}{1 + \exp(a + \gamma + b x)} + \lambda}.
The survival function is \eqn{S(x) = ((1 + \exp(a + \gamma)) / (1 + \exp(a + \gamma + b x)))^{\exp(-\gamma) / b} / \exp(\lambda x)}.
}
\examples{
dperks(80:100, c(-10, 0.09, -3, 0))
pperks(80:100, c(-10, 0.09, -3, 0))
qperks(seq(0, 1, by = 0.1), c(-10, 0.09, -3, 0))
rperks(10, c(-10, 0.09, -3, 0))
}
