% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-mortfit.R
\name{st_fit}
\alias{st_fit}
\title{Fit an LTRC survival tree (structure only)}
\usage{
st_fit(formula, data = NULL, data_train = NULL, data_test = NULL)
}
\arguments{
\item{formula}{A survival formula.}

\item{data, data_train, data_test}{See \code{baseline_fit()}.}
}
\value{
An \code{STModel} object.
}
\description{
Fit an LTRC survival tree (structure only)
}
