% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-mortfit.R
\name{stp_fit}
\alias{stp_fit}
\title{Fit a survival tree with per-node PAC models (STP)}
\usage{
stp_fit(
  formula,
  data = NULL,
  data_train = NULL,
  data_test = NULL,
  AC = "Beard"
)
}
\arguments{
\item{formula}{A survival formula.}

\item{data, data_train, data_test}{See \code{baseline_fit()}.}

\item{AC}{AC family or PAC object for node-wise fits.}
}
\value{
An \code{STPModel} object.
}
\description{
Fit a survival tree with per-node PAC models (STP)
}
