% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-fns.R
\name{survival_beard}
\alias{survival_beard}
\alias{cumulative_hazard_beard}
\alias{forceofmortality_beard}
\title{Beard mortality law}
\usage{
survival_beard(par, x)

cumulative_hazard_beard(par, x, t)

forceofmortality_beard(par, x)
}
\arguments{
\item{par}{numeric length-3 vector \code{c(a, b, gam)} where \code{gam = \gamma}.}

\item{x}{numeric vector of ages or durations since origin.}

\item{t}{numeric vector of time intervals; used by \code{cumulative_hazard_gompertz}.}
}
\description{
Logistic hazard with late-age leveling (Beard).
}
\details{
\eqn{\mu(x) = \frac{\exp(a + b x)}{1 + \exp(a + \gamma + b x)}}.
}
\examples{
survival_beard(c(-10, 0.09, -3), x = 80:100)
}
\seealso{
Other mortality-laws: 
\code{\link{survival_gompertz}()},
\code{\link{survival_kannisto}()},
\code{\link{survival_makeham}()},
\code{\link{survival_perks}()},
\code{\link{survival_thatcher}()}
}
\concept{mortality-laws}
