% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-fns.R
\name{survival_gompertz}
\alias{survival_gompertz}
\alias{cumulative_hazard_gompertz}
\alias{forceofmortality_gompertz}
\title{Gompertz mortality law}
\usage{
survival_gompertz(par, x)

cumulative_hazard_gompertz(par, x, t)

forceofmortality_gompertz(par, x)
}
\arguments{
\item{par}{numeric length-2 vector \code{c(a, b)}.}

\item{x}{numeric vector of ages or durations since origin.}

\item{t}{numeric vector of time intervals; used by \code{cumulative_hazard_gompertz}.}
}
\value{
For \code{survival_gompertz}: numeric vector of survival probabilities.

For \code{cumulative_hazard_gompertz}: numeric vector of cumulative hazards over \code{t}.

For \code{forceofmortality_gompertz}: numeric vector of hazards \eqn{\mu(x)}.
}
\description{
The Gompertz model assumes a hazard that grows exponentially with age.
}
\details{
Parameterization: \eqn{\mu(x) = \exp(a + b x)}.
The survival is \eqn{S(x) = \exp\{-\exp(a) (e^{b x} - 1)/b\}}.
}
\examples{
survival_gompertz(c(-10, 0.09), x = 80:100)
}
\seealso{
Other mortality-laws: 
\code{\link{survival_beard}()},
\code{\link{survival_kannisto}()},
\code{\link{survival_makeham}()},
\code{\link{survival_perks}()},
\code{\link{survival_thatcher}()}
}
\concept{mortality-laws}
