% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-fns.R
\name{survival_kannisto}
\alias{survival_kannisto}
\alias{cumulative_hazard_kannisto}
\alias{forceofmortality_kannisto}
\title{Kannisto mortality law}
\usage{
survival_kannisto(par, x)

cumulative_hazard_kannisto(par, x, t)

forceofmortality_kannisto(par, x)
}
\arguments{
\item{par}{numeric length-2 vector \code{c(a, b)}.}

\item{x}{numeric vector of ages or durations since origin.}

\item{t}{numeric vector of time intervals; used by \code{cumulative_hazard_gompertz}.}
}
\description{
Two-parameter logistic mortality with late-age deceleration.
}
\details{
\eqn{\mu(x) = \frac{\exp(a + b x)}{1 + \exp(a + b x)}}.
}
\examples{
survival_kannisto(c(-10, 0.09), x = 80:100)
}
\seealso{
Other mortality-laws: 
\code{\link{survival_beard}()},
\code{\link{survival_gompertz}()},
\code{\link{survival_makeham}()},
\code{\link{survival_perks}()},
\code{\link{survival_thatcher}()}
}
\concept{mortality-laws}
