% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-fns.R
\name{survival_makeham}
\alias{survival_makeham}
\alias{cumulative_hazard_makeham}
\alias{forceofmortality_makeham}
\title{Makeham mortality law}
\usage{
survival_makeham(par, x)

cumulative_hazard_makeham(par, x, t)

forceofmortality_makeham(par, x)
}
\arguments{
\item{par}{numeric length-3 vector \code{c(a, b, lambda)}.}

\item{x}{numeric vector of ages or durations since origin.}

\item{t}{numeric vector of time intervals; used by \code{cumulative_hazard_gompertz}.}
}
\description{
Gompertz with an age-independent background hazard \eqn{\lambda}.
}
\details{
\eqn{\mu(x) = \exp(a + b x) + \lambda}.
}
\examples{
survival_makeham(c(-10, 0.09, 0.0005), x = 80:100)
}
\seealso{
Other mortality-laws: 
\code{\link{survival_beard}()},
\code{\link{survival_gompertz}()},
\code{\link{survival_kannisto}()},
\code{\link{survival_perks}()},
\code{\link{survival_thatcher}()}
}
\concept{mortality-laws}
