% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-fns.R
\name{survival_perks}
\alias{survival_perks}
\alias{cumulative_hazard_perks}
\alias{forceofmortality_perks}
\title{Perks mortality law}
\usage{
survival_perks(par, x)

cumulative_hazard_perks(par, x, t)

forceofmortality_perks(par, x)
}
\arguments{
\item{par}{numeric length-4 vector \code{c(a, b, gam, lambda)} where \code{gam = \gamma}.}

\item{x}{numeric vector of ages or durations since origin.}

\item{t}{numeric vector of time intervals; used by \code{cumulative_hazard_gompertz}.}
}
\description{
Logistic hazard with late-age deceleration (Perks). Optional background \eqn{\lambda}.
}
\details{
\eqn{\mu(x) = \frac{\exp(a + b x)}{1 + \exp(a + \gamma + b x)} + \lambda}.
}
\examples{
survival_perks(c(-10, 0.09, -3, 0), x = 80:100)
}
\seealso{
Other mortality-laws: 
\code{\link{survival_beard}()},
\code{\link{survival_gompertz}()},
\code{\link{survival_kannisto}()},
\code{\link{survival_makeham}()},
\code{\link{survival_thatcher}()}
}
\concept{mortality-laws}
