% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-fns.R
\name{survival_thatcher}
\alias{survival_thatcher}
\alias{cumulative_hazard_thatcher}
\alias{forceofmortality_thatcher}
\title{Thatcher mortality law}
\usage{
survival_thatcher(par, x)

cumulative_hazard_thatcher(par, x, t)

forceofmortality_thatcher(par, x)
}
\arguments{
\item{par}{numeric length-3 vector \code{c(a, b, lambda)}.}

\item{x}{numeric vector of ages or durations since origin.}

\item{t}{numeric vector of time intervals; used by \code{cumulative_hazard_gompertz}.}
}
\description{
Kannisto with an age-independent background hazard \eqn{\lambda}.
}
\details{
\eqn{\mu(x) = \frac{\exp(a + b x)}{1 + \exp(a + b x)} + \lambda}.
}
\examples{
survival_thatcher(c(-10, 0.09, 0.0005), x = 80:100)
}
\seealso{
Other mortality-laws: 
\code{\link{survival_beard}()},
\code{\link{survival_gompertz}()},
\code{\link{survival_kannisto}()},
\code{\link{survival_makeham}()},
\code{\link{survival_perks}()}
}
\concept{mortality-laws}
