% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-fns.R
\name{Thatcher}
\alias{Thatcher}
\alias{dthatcher}
\alias{pthatcher}
\alias{qthatcher}
\alias{rthatcher}
\title{The Thatcher distribution}
\usage{
dthatcher(x, par)

pthatcher(x, par)

qthatcher(p, par, lower.tail = TRUE)

rthatcher(n, par)
}
\arguments{
\item{x, q}{numeric vector of quantiles.}

\item{par}{numeric length-3 vector of parameters \code{c(a, b, lambda)}.}

\item{p}{numeric vector of probabilities.}

\item{lower.tail}{logical; if \code{FALSE}, probabilities are \eqn{P(X > x)}.}

\item{n}{number of observations.}
}
\value{
\itemize{
\item \code{dthatcher()}: density.
\item \code{pthatcher()}: distribution function.
\item \code{qthatcher()}: quantiles.
\item \code{rthatcher()}: random values.
}
}
\description{
Density, distribution function, quantile function and random generation for the Thatcher distribution.
}
\details{
The Thatcher distribution is defined by the hazard function \eqn{\mu(x) = \frac{\exp(a + b x)}{1 + \exp(a + b x)} + \lambda}.
The survival function is \eqn{S(x) = ((1 + \exp(a)) / (1 + \exp(a + b x)))^{1 / b} / \exp(\lambda x)}.
}
\examples{
dthatcher(80:100, c(-10, 0.09, 0.0005))
pthatcher(80:100, c(-10, 0.09, 0.0005))
qthatcher(seq(0, 1, by = 0.1), c(-10, 0.09, 0.0005))
rthatcher(10, c(-10, 0.09, 0.0005))
}
