% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{uimm-package}
\alias{uimm}
\alias{uimm-package}
\title{uimm: An universe of individual level mortality models}
\description{
S4 classes and methods for parametric age-cohort (PAC) survival models (Gompertz, Makeham, Perks, Beard, Kannisto, Thatcher), with fitting, evaluation (AIC, IBS), a formula interface, survival trees with left-truncation, model-based recursive partitioning, and a Gompertz proportional hazards wrapper.
}
\author{
\strong{Maintainer}: Xiaochuan Lu \email{xiaochuan.lu@unsw.edu.au}

}
\keyword{internal}
